refactor: simplify frontend for backend development

+++ internal/middleware/auth.go
@@ -57,14 +57,15 @@ func (m *Middleware) AuthMiddleware() gin.HandlerFunc {
 			return
 		}
 
-		userid, ok := claims["id"].(int64)
+		//戻す際数字はデフォルトのfloat
+		userid, ok := claims["id"].(float64)
 		if !ok {
 			c.JSON(http.StatusUnauthorized, gin.H{"エラー": "ログイン中にエラーが起きました。"})
 			c.Abort()
 			return
 		}
 
-		c.Set("userID", userid)
+		c.Set("userID", int64(userid))
 		c.Set("username", username)
 		c.Next()
 	}
